pyproject.toml
[tool.poetry]
name = "pbthal-archive-manager"
version = "0.1.0"
description = ""
authors = ["Obscurely <adrian.obscurely@protonmail.com>"]
readme = "README.md"
packages = [{include = "pbthal_archive_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
music-tag = "^0.4.3"
gnureadline = "^8.1.2"
python-dotenv = "^0.21.0"
pillow = "^9.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
